ai_news_app/requirements.txt
feedparser
schedule
requests
python-dotenv
python-dateutil
deep-translator
flask
